+++ R/sql_node.R
@@ -4,7 +4,9 @@
 #'
 #' @param source source to rename from.
 #' @param exprs SQL expressions
+#' @param ... force later arguments to bind by name
 #' @param mods SQL modifiers (GROUP BY, ORDER BY, and so on)
+#' @param orig_columns logical if TRUE select all original columns.
 #' @return sql node.
 #'
 #' @examples
@@ -21,27 +23,42 @@
 #'
 #' @export
 #'
-sql_node <- function(source, exprs, mods = NULL) {
+sql_node <- function(source, exprs,
+                     ...,
+                     mods = NULL,
+                     orig_columns = TRUE) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "sql_node")
   UseMethod("sql_node", source)
 }
 
 #' @export
-sql_node.relop <- function(source, exprs, mods = NULL) {
+sql_node.relop <- function(source, exprs,
+                           ...,
+                           mods = NULL,
+                           orig_columns = TRUE) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "sql_node.relop")
   r <- list(source = list(source),
             table_name = NULL,
             parsed = NULL,
             exprs = exprs,
-            mods = mods)
+            mods = mods,
+            orig_columns = orig_columns)
   r <- relop_decorate("relop_sql", r)
   r
 }
 
 #' @export
-sql_node.data.frame <- function(source, exprs, mods = NULL) {
+sql_node.data.frame <- function(source, exprs,
+                                ...,
+                                mods = NULL,
+                                orig_columns = TRUE) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "sql_node.data.frame")
   tmp_name <- mk_tmp_name_source("rquery_tmp")()
   dnode <- table_source(tmp_name, colnames(source))
   dnode$data <- source
-  enode <- sql_node(dnode, exprs = exprs, mods = mods)
+  enode <- sql_node(dnode, exprs = exprs,
+                    mods = mods,
+                    orig_columns = orig_columns)
   return(enode)
 }
 
@@ -50,29 +67,31 @@ sql_node.data.frame <- function(source, exprs, mods = NULL) {
 
 #' @export
 column_names.relop_sql <- function (x, ...) {
-  if(length(list(...))>0) {
-    stop("unexpected arguments")
+  wrapr::stop_if_dot_args(substitute(list(...)), "column_names.relop_sql")
+  nms <- names(x$exprs)
+  if(x$orig_columns) {
+    nms <- c(nms, column_names(x$source[[1]]))
   }
-  names(x$exprs)
+  nms
 }
 
 
 
 #' @export
 format.relop_sql <- function(x, ...) {
-  if(length(list(...))>0) {
-    stop("unexpected arguments")
-  }
+  wrapr::stop_if_dot_args(substitute(list(...)), "format.relop_sql")
   assignments <- paste(names(x$exprs), ":=", as.character(x$exprs))
   modsstr <- ""
+  indent_sep <- "\n             "
   if(!is.null(x$mods)) {
     modsstr <- paste(";\n          ", x$mods)
   }
   paste0(trimws(format(x$source[[1]]), which = "right"),
          " %.>%\n ",
          "sql_node(.,\n",
-         "          ", paste(assignments, collapse = "\n             "),
+         "          ", paste(assignments, collapse = indent_sep),
          modsstr,
+         ",", indent_sep, "*=", x$orig_columns,
          ")\n")
 }
 
@@ -98,14 +117,15 @@ to_sql.relop_sql <- function (x,
                               tnum = mk_tmp_name_source('tsql'),
                               append_cr = TRUE,
                               using = NULL) {
-  if(length(list(...))>0) {
-    stop("unexpected arguments")
-  }
+  wrapr::stop_if_dot_args(substitute(list(...)), "to_sql.relop_sql")
   colsA <- vapply(names(x$exprs),
                   function(ci) {
                     quote_identifier(db, ci)
                   }, character(1))
   cols <- paste(as.character(x$exprs), "AS", colsA)
+  if(x$orig_columns) {
+    cols <- c("*", cols)
+  }
   subsql <- to_sql(x$source[[1]],
                    db = db,
                    source_limit = source_limit,
@@ -115,6 +135,7 @@ to_sql.relop_sql <- function (x,
                    using = NULL)
   tab <- tnum()
   prefix <- paste(rep(' ', indent_level), collapse = '')
+  star_str <- ""
   q <- paste0(prefix, "SELECT\n",
               prefix, " ", paste(cols, collapse = paste0(",\n", prefix, " ")), "\n",
               prefix, "FROM (\n",
@@ -130,9 +151,4 @@ to_sql.relop_sql <- function (x,
   q
 }
 
-#' @export
-#'
-dim.relop_sql <- function(x) {
-  c(NA_real_, NA_real_)
-}